refactor: Move fused layer norm funcs to layer_norm.py
